+++ activesupport/test/core_ext/time_ext_test.rb
@@ -430,6 +430,7 @@ class TimeExtCalculationsTest < ActiveSupport::TestCase
     assert_equal Time.local(2005, 1, 2, 11, 22, 33, 8), Time.local(2005, 1, 2, 11, 22, 33, 2).change(nsec: 8000)
     assert_raise(ArgumentError) { Time.local(2005, 1, 2, 11, 22, 33, 8).change(usec: 1, nsec: 1) }
     assert_nothing_raised { Time.new(2015, 5, 9, 10, 00, 00, "+03:00").change(nsec: 999999999) }
+    assert_raise(ArgumentError) { Time.local(2005, 1, 2, 11, 22, 33, 8).change(badarg: 42) }
   end
 
   def test_utc_change